backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
motor==3.6.0
python-dotenv==1.0.1
pydantic==2.9.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
slowapi==0.1.9